Changed prices to doubles

+++ src/main/java/maas/objects/Product.java
@@ -14,8 +14,8 @@ public class Product {
 	private int coolingRate;
 	private int boxingTemp;
 	private int breadsPerBox;
-	private int productionCost;
-	private int salesPrice;
+	private double productionCost;
+	private double salesPrice;
 	
 	public Product(String jsonProduct) {
 		// Parse the jsonProduct
@@ -30,8 +30,8 @@ public class Product {
 		this.coolingRate = product.getInt("cooling_rate");
 		this.boxingTemp = product.getInt("boxing_temp");
 		this.breadsPerBox = product.getInt("breads_per_box");
-		this.productionCost = product.getInt("production_cost");
-		this.salesPrice = product.getInt("sales_price");
+		this.productionCost = product.getDouble("production_cost");
+		this.salesPrice = product.getDouble("sales_price");
 	}
 	public String getId() {
 		return id;
@@ -63,10 +63,10 @@ public class Product {
 	public int getBreadsPerBox() {
 		return breadsPerBox;
 	}
-	public int getProductionCost() {
+	public double getProductionCost() {
 		return productionCost;
 	}
-	public int getSalesPrice() {
+	public double getSalesPrice() {
 		return salesPrice;
 	}
